requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
torch>=1.10,<2.0
channels>=3.0,<4.0
langchain
transformers
djangorestframework-simplejwt
langflow
rich
django-cors-headers
flask
